requirements.txt
beautifulsoup4==4.12.2
selenium==4.15.2
DrissionPage==4.1.0.17
python-dotenv==1.0.0
google-generativeai==0.3.1
uvicorn==0.25.0
pydantic>=2.11.3
requests==2.31.0
typing-extensions==4.12.0
starlette==0.29.0
jinja2==3.1.2
python-multipart==0.0.6
email-validator==2.1.0
httpx==0.24.1
ujson==5.10.0
websockets==12.0
aiofiles==23.2.1 
fastapi